% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_qmap.R
\name{core_qmap}
\alias{core_qmap}
\title{Correct predicted core values based on a known top-of-core value}
\usage{
core_qmap(
  core_mean,
  core_sd,
  core_top_val,
  lake_true_val,
  qmap_method = "delta",
  delta_transform = "multiplicative",
  interp_method = "linear",
  precision = 4,
  conf_alpha = 0.05,
  converge_accept = 1e-04,
  tempstep = NULL,
  iterations = NULL,
  seed = NULL,
  do_parallel = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{core_mean}{\strong{Numeric vector} of mean values in the cores to be corrected
(1 per core).}

\item{core_sd}{\strong{Numeric vector} of standard deviation values in the cores to
be corrected (1 per core).}

\item{core_top_val}{\strong{Numeric vector} of top-of-core values to be corrected (1
per core).}

\item{lake_true_val}{\strong{Numeric vector} of known values that the top-of-core
values should be matched to (1 per core).}

\item{qmap_method}{\strong{String.}
\itemize{
\item \code{"delta"} (default): Quantile mapping method similar to Quantile Delta
Mapping (QDM; Cannon et al., 2015)
\item \code{"empirical"}: Quantile mapping method similar to Empirical Quantile
Mapping (EQM)
\item \verb{"arithmetic}: Not true quantile mapping, but will return transformed
values based on an arithmetic calculation of an optimal value.
\item \code{"naive"}: Not true quantile mapping, but will return transformed values
based on a naive scaling of the original values.
}}

\item{delta_transform}{\strong{String.} Only used if \code{qmap_method = "delta"}.
\itemize{
\item \code{"multiplicative"} (default): Delta-m should be a multiplicative scalar.
Since the means and standard deviations should be greater than 0, this is
preferable.
\item \code{additive}: Delta-m should be an additive scalar.
}}

\item{interp_method}{\strong{String.} Any method that can be passed to
\code{\link[imputeTS:na_interpolation]{imputeTS::na_interpolation()}}. Controls how values are imputed for filling
out the distribution functions. Default is \code{"linear"}.}

\item{precision}{\strong{Integer.} Controls how many decimal places should be used
for interpolating distribution functions. Default is \code{4}.}

\item{conf_alpha}{\strong{Float.} The alpha value to use for calculating confidence
intervals. Default is \code{0.05}.}

\item{converge_accept}{\strong{Float.} The acceptance threshold for model convergence.
Refers to the minimum error between the modelled prediction and the known
value supplied. Default is \code{0.0001}.}

\item{tempstep}{\strong{Float.} Controls how quickly the temperature of the searching
algorithm decreases, which controls the narrowness of the searching
distribution. If not supplied, has different defaults depending on the method.
Default for \code{qmap_method = "delta"} is \code{0.0001}; default for
\code{qmap_method = "empirical"} is \code{0.001}.}

\item{iterations}{\strong{Integer.} Number of maximum iterations to search for optimal
weightings, if convergence is not achieved. If not supplied, has different
defaults depending on the method. Default for \code{qmap_method = "delta"} is
\code{10000}; default for \code{qmap_method = "empirical"} is \code{100}.}

\item{seed}{\strong{Integer.} For reproducibility, a seed to pass on to the searching
algorithm.}

\item{do_parallel}{\strong{Boolean.} If \code{TRUE} (default), samples will be searched
for optimal weights in parallel to speed up computation. If \code{FALSE}, samples
will be searched for optimal weights sequentially.}

\item{quiet}{\strong{Boolean.} If \code{FALSE} (default), information about weight
searching and potential issues will be printed to the console. If \code{TRUE},
nothing will be printed unless the function encounters a fatal error.}
}
\value{
An object of class \code{"coreqm"}, including:
\item{\code{opt_means}}{A vector of optimally transformed means, 1 per core.}
\item{\code{opt_means_ci}}{A data frame including the upper and lower confidence
intervals for the optimally transformed means.}
\item{\code{opt_sds}}{A vector of optimally transformed standard deviations, 1 per
core.}
\item{\code{opt_sds_ci}}{A data frame including the upper and lower confidence
intervals for the optimally transformed standard deviations.}
\item{\code{opt_weights}}{A data frame including the optimal weights for each core.}
\item{\code{internal_data}}{A data frame that includes the original data and
arithmetic and naive transformed values.}
\item{\code{orig_means_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of original
core means.}
\item{\code{orig_sds_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of original
core standard deviations.}
\item{\code{err_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of original
core errors.}
\item{\code{adj_means_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of
arithmetically adjusted core means.}
\item{\code{adj_sds_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of
arithmetically adjusted core standard deviations.}
\item{\code{naive_means_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of naively
adjusted core means.}
\item{\code{naive_sds_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of naively
adjusted core standard deviations.}
\item{\code{opt_means_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of optimally
adjusted core means.}
\item{\code{opt_sds_ecdf}}{An object of type \code{"qmrecdf"} for the ECDF of optimally
adjusted core standard deviations.}
\item{\code{weight_error_convergences}}{An object of type \code{"qdmerrmat"} that contains
information about the convergence of weights as the error function is
minimized.}
}
\description{
\code{core_qmap} adjusts predicted variables in lake core sediments using a
modification of quantile mapping, a climate modelling bias correction
technique, using the difference between the predicted value at the top
of the core and a known value at the top of the core, usually in water.
}
\details{
This function returns an object that can make adjustments to the
core values using \code{predict()}, but does not make the adjustments itself.
Ensure that the values used in \code{core_mean}, \code{core_sd}, \code{core_top_val}, and
\code{lake_true_val} align between cores.
}
