% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.qmrecdf}
\alias{plot.qmrecdf}
\title{Plot Interpolated ECDFs}
\usage{
\method{plot}{qmrecdf}(x, ci = TRUE, conf_alpha = 0.05, col = "blue", ...)
}
\arguments{
\item{x}{An object of class \code{"qmrecdf"}.}

\item{ci}{\strong{Boolean.} If \code{TRUE} (default), plot C.I. around ECDF values.}

\item{conf_alpha}{\strong{Float.} The alpha value to use for calculating confidence
intervals. Default is \code{0.05}.}

\item{col}{\strong{String.} Colour to plot confidence interval lines.}

\item{...}{Further arguments to pass to \code{base::plot()}.}
}
\description{
Plot method for class \code{"qmrecdf"}. Capable of plotting the interpolated
ECDF along with a confidence interval.
}
