% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.qmrerrmat}
\alias{plot.qmrerrmat}
\title{Plot Convergence-Error values for Simulated Annealing Results}
\usage{
\method{plot}{qmrerrmat}(x, best_col = "red", try_col = "blue", ...)
}
\arguments{
\item{x}{An object of class \code{"qmrerrmat"}.}

\item{best_col}{\strong{String.} Colour for the best values.}

\item{try_col}{\strong{String.} Colour for the tried (neighbour) values.}

\item{...}{Further arguments to be passed to \code{base::plot()}.}
}
\description{
Plot method for objects of class \code{"qmrerrmat"}.
}
