% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_validate_coreqm.R
\name{predict.coreqm}
\alias{predict.coreqm}
\title{Whole-core predictions using a \code{coreqm} object}
\usage{
\method{predict}{coreqm}(
  object,
  core_vals,
  pb_dates = NULL,
  date_col = NULL,
  conf_alpha = 0.05,
  quiet = FALSE,
  clamp_zeroes = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of type \code{"coreqm"}. See \code{core_qmap()}, which
creates these objects.}

\item{core_vals}{\strong{A data frame} with at least 3 columns (in order):
\enumerate{
\item Column of site/sample IDs.
\item Column of midpoint depths for intervals.
\item Column of original predicted values for each interval.
\item Column of 210Pb dates for each interval (optional).
Any columns beyond these first 4 will be ignored.
}}

\item{pb_dates}{\strong{A data frame} (optional) that contains a 210Pb dating
profile for each core to be adjusted. If supplied, must contain at least 3
columns (in order):
\enumerate{
\item Column of site/sample IDs to match with \code{core_vals}.
\item Column of interval top/bottom/midpoint depths.
\item Column of inferred 210Pb ages by interval depth.
Any columns beyond the first 3 will be ignored.
}}

\item{date_col}{\strong{String.} (optional) If supplying a 4th column to \code{core_vals},
the name of the column containing 210Pb-inferred dates.}

\item{conf_alpha}{\strong{Float.} The alpha value to use for calculating confidence
intervals. Default is \code{0.05}.}

\item{quiet}{\strong{Boolean.} If \code{FALSE} (default), information about weight
searching and potential issues will be printed to the console. If \code{TRUE},
nothing will be printed unless the function encounters a fatal error.}

\item{clamp_zeroes}{\strong{Boolean.} Transforming inferred core values in this
way can sometimes result in downcore values < 0. If \code{TRUE} (default), any
value below 0 will be clamped to 0. If \code{FALSE}, no clamping is performed.}

\item{...}{Unused.}
}
\value{
A data frame based on \code{core_vals} with the following additional
columns:
\item{transformed_val}{Transformed values of the input measure.}
\item{.ci_lower_rng}{Lower range-based C.I.}
\item{.ci_upper_rng}{Upper range-based C.I.}
\item{.ci_lower_prb}{Lower distribution-based C.I.}
\item{.ci_upper_prb}{Upper distribution-based C.I.}
}
\description{
This predict method will apply the adjustments from a \code{coreqm} object to
transform downcore values based on the top-of-core adjustments. Variance
flattening using 210Pb-inferred dates is also available.
}
\details{
If either \code{pb_dates} or \code{date_col} is (correctly) supplied, the variances
of downcore measurements will be flattened according to the number of years
contained in each interval in comparison with the first interval. If \code{pb_dates}
is supplied, the flattening will be done according to the full dating profile
provided. If \code{date_col} is supplied, with a properly matching column in
\code{core_vals}, interval widths will be inferred from the modulus of the interval
midpoint, and dates will be linearly interpolated up and down to get the date
range of the interval.
If no variance flattening is desired, leave both \code{pb_dates} and \code{date_col}
as \code{NULL}.
}
