% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_validate_coreqm.R
\name{validate_coreqm}
\alias{validate_coreqm}
\title{Validate a coreqm object}
\usage{
validate_coreqm(
  coreqm_obj,
  interp_method = "linear",
  precision = 4,
  conf_alpha = 0.05,
  ...
)
}
\arguments{
\item{coreqm_obj}{An object of type \code{coreqm}.}

\item{interp_method}{\strong{String.} Any method that can be passed to
\code{\link[imputeTS:na_interpolation]{imputeTS::na_interpolation()}}. Controls how values are imputed for filling
out the distribution functions. Default is \code{"linear"}.}

\item{precision}{\strong{Integer.} Controls how many decimal places should be used
for interpolating distribution functions. Default is \code{4}.}

\item{conf_alpha}{\strong{Float.} The alpha value to use for calculating confidence
intervals. Default is \code{0.05}.}

\item{...}{Unused.}
}
\value{
A data frame with the following columns:
\item{core_top_val}{The original top-of-core values.}
\item{true_val}{The known values.}
\item{transformed_top_val}{Modelled top-of-core values.}
\item{.ci_lower_rng}{Lower range-based C.I.}
\item{.ci_upper_rng}{Upper range-based C.I.}
\item{.ci_lower_prb}{Lower distribution-based C.I.}
\item{.ci_upper_prb}{Upper distribution-based C.I.}
}
\description{
\code{validate_coreqm} will return the optimal transformed top-of-core values
(not downcore) for each sample so that the results can be verified.
}
